(feat) web worker to handle the timer ⏰
Goggle chrome tries to be a smart arse and slows the execution of
timers in an inactive tab, not ideal when trying to create a pomodoro
timer. One solution is to use a web worker, these are separate threads
that run outside of the UI thread (main thread in JavaSript).